Se trabajo en la parte de agregar productos duplicados  al carrito en la base de datos y calcular el total y la cantidad de un producto en el carrito

+++ database/models/Carrito.js
@@ -13,7 +13,7 @@ module.exports = (sequelize, dataTypes) => {
       allowNull: false
     },
     total: {
-      type: dataTypes.FLOAT,
+      type: dataTypes.DOUBLE,
       allowNull: false
     },
     product_id: {
